+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add claiming to organization integration details. Refs UIORGS-442.
 * Add "Duplicate" integration action to organization integration view. Refs UIORGS-441.
 * *BREAKING* Add number generator for vendor code including settings page. Refs UIORGS-336, UIORGS-337.
+* Enhancement help text on Settings > Organizations > Number generator options. Refs UIORGS-453.
 
 ## [5.2.0](https://github.com/folio-org/ui-organizations/tree/v5.2.0) (2024-10-31)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v5.1.1...v5.2.0)